Moving CardStatus out of storage code
