+++ todo_app/data/mongodb.py
@@ -7,23 +7,7 @@ import pymongo
 from bson.objectid import ObjectId
 
 from todo_app.data.item import Item
-
-
-class CardStatus(Enum):
-    DONE = 'Done'
-    TODO = 'To Do'
-    DOING = 'Doing'
-
-    @classmethod
-    def get_status(cls, status: str) -> Enum:
-        if status == cls.DONE.value:
-            return cls.DONE
-        elif status == cls.TODO.value:
-            return cls.TODO
-        elif status == cls.DOING.value:
-            return cls.DOING
-        else:
-            raise ValueError('Unknown status')
+from todo_app.data.card_status import CardStatus
 
 
 class MongoDB:
